refactor: Pet Impl 삭제 및 로직 수정
